requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.14.0
django-tinymce==3.5.0
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.3
